# NGSIv2 rendering improvements (simplification, unification and performance)