--- a/src/lib/common/limits.h
+++ b/src/lib/common/limits.h
@@ -211,12 +211,4 @@
 
 
 
-/* ****************************************************************************
-*
-* Precision constants -
-*/
-#define PRECISION_DIGITS  9
-#define PRECISION         0.000000001  // it corresponds to 9 digits
-
-
 #endif  // SRC_LIB_COMMON_LIMITS_H_