… src/lib/common/JsonHelper.h
… src/lib/common/macroSubstitute.cpp
… src/lib/common/string.cpp
… src/lib/common/string.h
… src/lib/common/tag.cpp
… src/lib/logMsg/time.h
… src/lib/mongoBackend/MongoCommonUpdate.cpp
… src/lib/mongoBackend/MongoGlobal.cpp
… src/lib/ngsi/ContextAttribute.cpp
… src/lib/ngsi/ContextAttribute.h
… src/lib/ngsi/ContextAttributeVector.cpp
… src/lib/ngsi/ContextAttributeVector.h
… src/lib/ngsi/ContextElement.cpp
… src/lib/ngsi/ContextElement.h
… src/lib/ngsi/ContextElementResponse.cpp
… src/lib/ngsi/ContextElementResponse.h
… src/lib/ngsi/ContextElementResponseVector.cpp
… src/lib/ngsi/ContextElementResponseVector.h
… src/lib/ngsi/Metadata.cpp
… src/lib/ngsi/Metadata.h
… src/lib/ngsi/MetadataVector.cpp
… src/lib/ngsi/MetadataVector.h
… src/lib/ngsi10/NotifyContextRequest.cpp
… src/lib/ngsi10/NotifyContextRequest.h
… src/lib/ngsiNotify/Notifier.cpp
… src/lib/ngsiNotify/Notifier.h
… src/lib/ngsiNotify/QueueNotifier.cpp
… src/lib/ngsiNotify/QueueNotifier.h
… src/lib/orionTypes/EntityType.cpp
… src/lib/orionTypes/EntityTypeResponse.cpp