This PR is motivated by https://github.com/telefonicaid/fiware-orion/issues/1298#issuecomment-417659196 and introduces improvements in NGSIv2 rendering logic in both simplfication/unification and peformance aspects. A summary of changes follow.

With regards to simplification/unification:

* Extend JsonHelper utilization along NGSIv2 rendering logic
* Simplification in several methods, mainly:
   * CompoundValueNode::toJson()
   * Entity::toJson() (spliting in several other methods)
   * ContextAttribute::toJson()
   * ContextAttributeVector:.toJson() (removed)
   * MetadataVector::toJson() (the usage of JsonHelper removes the need of double-pass)
* Simplification in argument passing. In particular, `bool comma` removed whenever possible
* Removing stale code.

With regards to peformance:

* JsonHelper implemention changed from std::ostringstream to std::string concatenation (as in previous NGSIv2 renderind code for entities). After the changes in this PR, I think std::ostringstream is no longer used in any part of the CB code.
* Logic in the toJson() methods has been simplified.
* I have done some basic `ab` testing with render intenstive operations (GET /v2/entities?limit=1000 and GET /v2/subscriptions?limit=1000, which 500 bytes entities/subscription, i.e. ~50KB in each GET response). The details of the test are [here for master](https://gist.github.com/fgalan/cdde23007ed7b34b566be72c67121907) and [here for this branch](https://gist.github.com/fgalan/d84d75ec0a66f53672257c860525c2df), but, in summary the result is pretty good:
    * Entities rendering test: from 77.22 tps to 103.06 tps **(33.46% gain)**
    * Subscriptions renderint test: 25.73 tps from to 119.75 tps **(365.4% gain)**

(The script used to populate the database can be found [here](https://gist.github.com/fgalan/e65be5ea398f8553b6a331885f2872e4). Eventually, it would be added to the repository for future usage)

From my point of view, this PR doesn't completes #1298 and more could to be done with regards to simplification and unification (for instance, ContextElement should use Entity internally and this will simplify a lot removing all the specific ContextElement rendering code) and peformance (other implementation alternatives for JsonHelper.cpp could be explored) but I think is an important step ahead.

Note, there are some changes in .test but all there seem to be ok. See specific comments in .test files.