@kzangeli 2018-09-06
LGTM (assuming valgrind reports no leaks)