@fgalan 2018-09-07
Some additional changes were required after LGTM, due to memory management issues. The new changes can be seen at https://github.com/telefonicaid/fiware-orion/compare/6eafb3f...7a31172